// backend/pdf-ranking-service/src/cv/interfaces/cv-analysis.interface.ts
export interface CVAnalysisResult {
  name: string;
  email: string;
  city: string;
  age: number;
  university: string;
  score: number;
}

export interface ApplicantData {
  name: string;
  email: string;
  cv_link: string;
  age: number;
  campaign_id: string;
  city: string;
  university: string;
  score: number;
}